chore: rearrange some files to make ESLint work
somehow the tsconfig.node.json secondary config that is only used when running the project locally in dev mode is not loaded properly by ESLint and it throws errors about `strictNullChecks` not being set to true.

I don't have time to look into this, so I'll just simplify the configs.